Temporary workaround for number keyboard issue
Due to an issue with the number pad keyboard not correctly appearing, a temporary workaround has been implemented by enabling options for signed numbers and adding a filter to only accept digit inputs. This change applies to the inputs of both the 'reps' and 'weight' fields on the new workout screen. This is a temporary solution until the issue with the number keyboard is fixed.