Add dialog for muscle group selection in searchable dropdown
In the searchable dropdown widget, a new feature was added where, if the muscle group related to the search item is not defined, a dialog is set to prompt the user to select the primary and secondary muscle groups. This is done using the "_promptMuscleGroup" function that calls a new widget, "SelectMuscleGroupDialog". This change improves the overall user experience by providing a more interactive way to specify muscle groups related to a search term.