Update petitparser version, change logo image and remove unused import
Updated the petitparser package from version 6.0.1 to 6.0.2 to get the newest features. On the login screen, the application's logo image reference was changed from 'fitness_app_logo.png' to 'peak_logo.png', to reflect the new branding applied recently. Also, an unnecessary import was removed in searchable_dropdown.dart, which helps in simplifying the codebase and improving readability. Lastly, the app icon reference in pubspec.yaml was updated to 'peak_logo_background.png' for both iOS and Android, emphasizing the branding changes.