Update SearchableDropdown to be fullscreen modal
The custom SearchableDropdown widget has been updated to not just change the background color, but to also make the dropdown expandable to a full-screen modal with a draggable functionality. The initial size of the modal is 80% of the full screen and it can be dragged to full size or minimized to half-size. This change was made to enhance the user experience by making the list of options more visible and interactive.