Creation of custom exercises (#42)
Users need to be able to create custom exercises and save them to their
firestore account. Other users in the app shouldn't be able to see other
users custom exercises. After creating the custom exercise, users should
be able to search for it during a workout to pull in that exercise.

Also fixed a last login update issue. If the user opened up the app
again and was already logged in, we never updated the last login
property in firebase.