Features and expectations (#49)
This PR updates our README with a clear definition of what this app
needs to be capable of doing in the future. Also this PR updates and
creates models that the system will need to be able to handle. Future
PRs will need to resolve the current code and integrate it with the
models. Going to make this PR just to get the new models into the repo
and start focusing on cleaning up other code sections in other issues
and PRs.

Resolves #44